// src/service/member/schemas/followResult.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * Member API
 * isd / kinder board api doc
 * OpenAPI spec version: v1.1.0
 */
import type { ProfileResult } from './profileResult';

export interface FollowResult {
  /** 팔로잉 프로필 ID */
  followingProfileId: number;
  /** 팔로워 프로필 ID */
  followerProfileId: number;
  /** 맞팔 여부 */
  isFollowBack?: boolean;
  /** 수정일 */
  modifiedAt?: string;
  /** 팔로우 날짜 */
  createdAt?: string;
  profile?: ProfileResult;
}